apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "python-dotenv>=1.1.1",
]
